데이터 녹화/재생 프로토타입 작업

+++ Assets/Scripts/Unit/Interface/IObject.cs
@@ -1,4 +1,4 @@
 public interface IObject
 {
-    public ulong id { get; set; }    
+    public int id { get; set; }    
 }
